refactor(sdk): consolidate client implementation and remove wrapper layer
